feat: make transactions soft-deletable through the ui
